[fbsync] Fix issue in label Transform (#1455)
Summary: In the construction of Vocab within label transform, the default index is set to 0. This index is returned when OOV token is given. For this transform, the default index should never be set. Otherwise, it will return default index (which is 0) for unknown labels that might get passed (Ideally it should throw error in this case because we do not know what to do when wrong label is passed for query)

Reviewed By: hudeven

Differential Revision: D32610834

fbshipit-source-id: e49385fb313929627c41fc515b6d900a6bfc3591