Revert D44660997: Multisect successfully blamed D44660997 for test or build failures
Summary:
This diff is reverting D44660997
D44660997: [Codemod][PyTorchTextGithubSync] Import TorchText from Github by generatedunixname89002005287564 has been identified to be causing the following test or build failures:

Tests affected:
- [fblearner/flow/projects/pytorch/recipes/fb/projects/assistant/smart_replies/tests:test_workflow_needed_coverage - main](https://www.internalfb.com/intern/test/281475066660764/)

Here's the Multisect link:
https://www.internalfb.com/multisect/1879479
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: joecummings

Differential Revision: D44969435

fbshipit-source-id: 0b38ef6e60ebc30528d0d38ce6d15e8d82c69b82