add compat module to point fbcode users to torchdata_deprecated
Summary:
We're deleting datapipes in torchdata. TorchText ahs a bunch of datapipes defined, and we don't want to break fbcode tests. Internally we are moving users to `torchdata_deprecated` but we need to jump through some hoops.

This diff creates a "compat" module which redirects to torchdata.datapipes.iter by default, but in FBCode we patch it to point to torchdata_deprecated instead.

Reviewed By: ivy-zhou

Differential Revision: D64208916

fbshipit-source-id: 6b625d87f3c42b28207bd84de3417cc3bc74caee